refactor(package): move to domain package
