+++ README.md
@@ -95,6 +95,6 @@
 
 ### 구현할 기능 목록
 
-- [ ] Domain 패키지로 분리
+- [x] Domain 패키지로 분리
 - [ ] MVC 패턴 형식으로 분리
