+++ src/main/kotlin/study/racingcar/domain/car/Car.kt
@@ -1,4 +1,4 @@
-package study.racingcar.car
+package study.racingcar.domain.car
 
 class Car(val name: String) {
     init {
