+++ src/main/kotlin/study/racingcar/domain/car/CarFactory.kt
@@ -1,4 +1,4 @@
-package study.racingcar.car
+package study.racingcar.domain.car
 
 object CarFactory {
     fun createCars(names: List<String>): List<Car> {
