+++ src/main/kotlin/study/racingcar/domain/car/Power.kt
@@ -1,4 +1,4 @@
-package study.racingcar.car
+package study.racingcar.domain.car
 
 @JvmInline
 value class Power(val value: Int)
