+++ src/main/kotlin/study/racingcar/domain/racing/Racing.kt
@@ -1,8 +1,8 @@
-package study.racingcar.racing
+package study.racingcar.domain.racing
 
-import study.racingcar.car.Car
-import study.racingcar.car.Power
-import study.racingcar.generator.Generator
+import study.racingcar.domain.car.Car
+import study.racingcar.domain.car.Power
+import study.racingcar.domain.generator.Generator
 
 class Racing(private val cars: List<Car>, private val generator: Generator) {
     fun round(): List<Record> {
