+++ src/main/kotlin/study/racingcar/domain/racing/RacingResult.kt
@@ -1,3 +1,3 @@
-package study.racingcar.racing
+package study.racingcar.domain.racing
 
 data class RacingResult(val rounds: List<List<Record>>)
