+++ src/main/kotlin/study/racingcar/domain/racing/Record.kt
@@ -1,6 +1,6 @@
-package study.racingcar.racing
+package study.racingcar.domain.racing
 
-import study.racingcar.car.Car
+import study.racingcar.domain.car.Car
 
 data class Record(val name: String, val location: Int) {
     constructor(car: Car) : this(car.name, car.currentLocation)
