+++ src/main/kotlin/study/racingcar/domain/racing/Winner.kt
@@ -1,6 +1,6 @@
-package study.racingcar.racing
+package study.racingcar.domain.racing
 
-import study.racingcar.car.Car
+import study.racingcar.domain.car.Car
 
 class Winner(private val cars: List<Car>) {
     fun findWinners(): List<Car> {
