+++ src/test/kotlin/study/racingcar/domain/racing/RecordKoTest.kt
@@ -1,8 +1,8 @@
-package study.racingcar.racing
+package study.racingcar.domain.racing
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import study.racingcar.car.Car
+import study.racingcar.domain.car.Car
 
 internal class RecordKoTest : StringSpec({
     "Car만을 통해서 Record를 생성할 수 있따" {
